+++ lib/src/cli/commands/analyze.dart
@@ -34,6 +34,7 @@ class AnalyzeCommand extends BaseCommand {
   @override
   void validateCommand() {
     validateRootFolderExist();
+    validateSdkPath();
     validateTargetDirectories();
   }
 
@@ -54,6 +55,7 @@ class AnalyzeCommand extends BaseCommand {
       argResults.rest,
       argResults[FlagNames.rootFolder] as String,
       config,
+      findSdkPath(),
     );
 
     await _analyzer
@@ -92,6 +94,7 @@ class AnalyzeCommand extends BaseCommand {
     _usesReporterOption();
     _usesMetricsThresholdOptions();
     usesRootFolderOption();
+    usesSdkPathOption();
     usesExcludeOption();
     _usesExitOption();
   }