+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 * Introduce `check-unused-l10n` command.
+* Don't trigger prefer-extracting-callbacks on empty function blocks.
+* Improve unused files check, add support for `vm:entry-point` annotation.
 
 ## 4.3.1